+++ app/src/main/java/com/cswala/cswala/Splashscreen.java
@@ -1,4 +1,5 @@
-package com.cswala.cswala.Common;
+package com.cswala.cswala;
+
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
@@ -9,7 +10,6 @@ import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.cswala.cswala.R;
 import com.cswala.cswala.utils.IntentHelper;
 
 import com.pixplicity.easyprefs.library.Prefs;
