+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
-    tools:context=".Common.MainActivity">
+    tools:context=".MainActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
